+++ RentalSystem/RentalSystem/Entities/Vehicles/CargoVan.cs
@@ -2,7 +2,7 @@
 {
     public class CargoVan : Vehicle
     {
-        public CargoVan(string make, string model, float valuedAt) : base(make, model, valuedAt)
+        public CargoVan(string make, string model, decimal valuedAt) : base(make, model, valuedAt)
         {
         }
     }
